Meditation Audio - Custom - Python
Used:
scipy soundfile
imageio_ffmpeg & ffprobe from pydub
AudioSegment
subprocess
numpy

######
Soundfile could not read the sound clips that i downloaded from freesound.org (MP3 files) directly. I had to convert the MP3s to WAV first, then work with them. The sound clips were longer in durations then what i needed, so i had to write a script (python) to reduce the length. Github would not allow me to upload the file because the WAV file was too large because it was uncompressed audio. I converted it MP3, which was much smaller, but it still failed to be under the size limit (25MB) for Github, so i wrote another python script to reduce the bitrate to get it under 25MB. It worked. Now, I will create another medication clip that includes guided AUM chanting and the invocation chant at the beginning and the end of the practice for Shambhavi Mahamudra.  Extremely fun project.